BL-15 create helpers file inside the component and move helper functions to it
